Add files via upload
I've written this code to access a local MySQL database to display the data in a window. Database name, table name, column names, and MySQL password can be changed. 
Not the end solution, but its a step in the right direction

NOTE: Line 25 on JDBC.java there is a string "YES", this string must be changed to your local MySQL password in order to properly connect.